Upgrade the dependencies that don't require a migration
Splits the safely-upgradable subset out of the Dependabot group PR (#683),
which bundled them with hyper 0.14 -> 1.x, wasmtime 39 -> 47 and the
rustls 0.21 -> 0.23 stack. Those three need real source migrations and are
left for separate PRs.

No source changes needed:

  base64           0.21.2 -> 0.22.1
  itertools        0.10.5 -> 0.14.0   (also dedupes the two copies in the lock)
  thiserror        1.0.37 -> 2.0.18
  bytesize         1.1.0  -> 2.7.0
  proptest-derive  0.5.1  -> 0.8.0
  tls-listener     0.7    -> 0.8
  wasmparser  \
  wasm-encoder |   0.250  -> 0.256
  wit-component|
  wasm-metadata/

wasm-metadata moves in lockstep with the rest of the wasm-tools crates.
Leaving it behind means `adapt_bytes` re-parses a component encoded by
wit-component 0.256 using wasmparser 0.250, which fails at runtime rather
than at build time.

tls-listener 0.8 is still built against hyper 0.14 and tokio-rustls 0.24,
so it is only compatible while those stay put. It will need >= 0.10 as part
of the hyper 1.x migration.

Small migrations:

  walrus                0.23.3 -> 0.26.4
  rustls-native-certs   0.6.3  -> 0.8.4

walrus 0.26 widens `MemArg::offset` to u64 for memory64. rustls-native-certs
0.8 returns a `CertificateResult` rather than a `Result`, so load errors are
now reported as warnings; the existing empty-root-store check still turns a
total failure into `Error::TlsNoCAAvailable`.

Verified with `cargo check --workspace --all-targets` (clean) and
`cargo test --all` (309 passed, 0 failed).

+++ src/shift_mem.rs
@@ -143,7 +143,7 @@ fn shift_func(r#gen: &mut ModuleLocals, func: &mut LocalFunction) {
                     // avoid the compiler crash, in case this load instruction is unreachable.
                     // If the code is reachable, the original and adapted code will both crash,
                     // so there is no observable divergence.
-                    let offset = offset.saturating_add(OFFSET as u32);
+                    let offset = offset.saturating_add(OFFSET as u64);
                     let instr = Instr::Load(Load {
                         memory,
                         kind,
@@ -156,7 +156,7 @@ fn shift_func(r#gen: &mut ModuleLocals, func: &mut LocalFunction) {
                     kind,
                     arg: MemArg { align, offset },
                 }) => {
-                    let offset = offset.saturating_add(OFFSET as u32);
+                    let offset = offset.saturating_add(OFFSET as u64);
                     let instr = Instr::Store(Store {
                         memory,
                         kind,
@@ -169,7 +169,7 @@ fn shift_func(r#gen: &mut ModuleLocals, func: &mut LocalFunction) {
                     kind,
                     arg: MemArg { align, offset },
                 }) => {
-                    let offset = offset.saturating_add(OFFSET as u32);
+                    let offset = offset.saturating_add(OFFSET as u64);
                     let instr = Instr::LoadSimd(LoadSimd {
                         memory,
                         kind,
